It Works (Mostly)
A toy language implementation that is written in Rust that currently has
variable declarations, basic type inference, and simple binary
operators.

None of these are robust and will cause panics when provided invalid
input.

It uses LLVM for the codegen backend (and a fair amount of unsafe code).

Currently no executables or objects are output and only LLVM IR is
written to the screen.

Signed-off-by: Jacob Evans <jacob@jwevans.dev>