Start again to refactor and add tests
Remove existing project files and begin writing them again with lessons
learned from the first attempt. This includes adding tests as the
changes are made to the corresponding implementation files. In addition
to test the interfaces will be made more modular and have documentation
added to allow for easier refactoring in the future.

This commit adds a new `ZoidLocation` struct that will be used to help
add diagnostics to any errors or warning that are generated. It also
includes the beginning of the `ZoidLexer` that has the different kinds
of brackets that will be allowed. It adds `<:` and `:>` that may be used
if Zoid eventually supports generics.

Signed-off-by: Jacob Evans <jacob@jwevans.dev>